Move codegen code over
